Changed search forms positioning and clearing data functionality

+++ scripts/Controllers/UserController.js
@@ -22,9 +22,12 @@ class UserController{
     }
 
     loginUser(event){
-        event.preventDefault();
-        sessionStorage.clear();
-        Kinvey.User.logout();
+        event.preventDefault();  
+        if(sessionStorage.getItem('authToken')) {
+            sessionStorage.clear();
+            Kinvey.User.logout();     
+        }
+        
         let userObject = this.view.getDataFromLogin();
         let promise = Kinvey.User.login(userObject.username, userObject.password)
             .then(function(data){
